+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a13 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- Fix date format in collection tiles (closes `#584`_).
+  [tcurvelo]
+
 - Package is now compatible with plone.app.widgets as RichText tile no longer breaks with it (closes `#543`_).
   [frapell, rodfersou]
 
@@ -731,4 +734,5 @@ There's a frood who really knows where his towel is.
 .. _`#534`: https://github.com/collective/collective.cover/issues/534
 .. _`#543`: https://github.com/collective/collective.cover/issues/543
 .. _`#559`: https://github.com/collective/collective.cover/issues/559
+.. _`#584`: https://github.com/collective/collective.cover/issues/584
 .. _`PloneFormGen`: https://pypi.python.org/pypi/Products.PloneFormGen